refactor field solvers
- add all solver to the namespace `picongpu::fields::maxwellSolver`
- add the current interpolation as template parameter
- make the solver cell type aware
- provide forwared definitions in `*.def` files
