+++ include/picongpu/fields/MaxwellSolver/Yee/Yee.def
@@ -1,0 +1,82 @@
+/* Copyright 2013-2018 Axel Huebl, Heiko Burau, Rene Widera
+ *
+ * This file is part of PIConGPU.
+ *
+ * PIConGPU is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * PIConGPU is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with PIConGPU.
+ * If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+
+#include "picongpu/fields/MaxwellSolver/Yee/Curl.def"
+#include "picongpu/fields/currentInterpolation/CurrentInterpolation.def"
+
+
+namespace picongpu
+{
+namespace fields
+{
+namespace maxwellSolver
+{
+
+    template<
+        typename T_CurrentInterpolation = currentInterpolation::None,
+        typename CurlE = yee::CurlRight,
+        typename CurlB = yee::CurlLeft
+    >
+    class Yee;
+
+} // namespace maxwellSolver
+} // namespace fields
+
+namespace traits
+{
+
+    template<
+        typename T_CurrentInterpolation,
+        class CurlE,
+        class CurlB
+    >
+    struct GetMargin<
+        picongpu::fields::maxwellSolver::Yee<
+            T_CurrentInterpolation,
+            CurlE,
+            CurlB
+        >, FIELD_B
+    >
+    {
+        using LowerMargin = typename CurlB::LowerMargin;
+        using UpperMargin = typename CurlB::UpperMargin;
+    };
+
+    template<
+        typename T_CurrentInterpolation,
+        class CurlE,
+        class CurlB
+    >
+    struct GetMargin<
+        picongpu::fields::maxwellSolver::Yee<
+            T_CurrentInterpolation,
+            CurlE,
+            CurlB
+        >,
+        FIELD_E
+    >
+    {
+        using LowerMargin = typename CurlE::LowerMargin;
+        using UpperMargin = typename CurlE::UpperMargin;
+    };
+
+} //namespace traits
+} // namespace picongpu
