--- include/picongpu/fields/MaxwellSolver/Yee/YeeSolver.def
@@ -1,59 +1,0 @@
-/* Copyright 2013-2018 Axel Huebl, Heiko Burau, Rene Widera
- *
- * This file is part of PIConGPU.
- *
- * PIConGPU is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * PIConGPU is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with PIConGPU.
- * If not, see <http://www.gnu.org/licenses/>.
- */
-
-
-
-#pragma once
-
-#include "picongpu/fields/MaxwellSolver/Yee/Curl.hpp"
-#include "picongpu/algorithms/DifferenceToUpper.hpp"
-#include "picongpu/algorithms/DifferenceToLower.hpp"
-
-namespace picongpu
-{
-namespace yeeSolver
-{
-using CurlLeft = Curl< DifferenceToLower< simDim > >;
-using CurlRight = Curl< DifferenceToUpper< simDim > >;
-
-template<class CurlE = CurlRight, class CurlB = CurlLeft>
-class YeeSolver;
-} // namespace yeeSolver
-
-
-namespace traits
-{
-
-template<class CurlE, class CurlB>
-struct GetMargin<picongpu::yeeSolver::YeeSolver<CurlE, CurlB>, FIELD_B>
-{
-    using LowerMargin = typename CurlB::LowerMargin;
-    using UpperMargin = typename CurlB::UpperMargin;
-};
-
-template<class CurlE, class CurlB>
-struct GetMargin<picongpu::yeeSolver::YeeSolver<CurlE, CurlB>, FIELD_E>
-{
-    using LowerMargin = typename CurlE::LowerMargin;
-    using UpperMargin = typename CurlE::UpperMargin;
-};
-
-} //namespace traits
-
-} // namespace picongpu
